Convert the integer to number

+++ src/store/table/tableReducer.js
@@ -187,8 +187,8 @@ export const reducers = {
     }
 
     switch (action.dataType) {
-      case "integer":
-        if (state.columns[typeIndex].dataType === "integer") {
+      case "numeric":
+        if (state.columns[typeIndex].dataType === "numeric") {
           return state;
         } else {
           return {
@@ -202,7 +202,7 @@ export const reducers = {
               ...row,
               [action.columnId]: isNaN(row[action.columnId])
                 ? ""
-                : Number.parseInt(row[action.columnId])
+                : Number.parseFloat(row[action.columnId])
             }))
           };
         }
